Functionality added
I added links to the buttons, some don't work because the route doesn't exist yet. There is also no functionality for disabling buttons based on permissions yet. I also didn't know what to do for the "Other" section in terms of how it should work so I will leave it blank for now.